+++ library/BOTK/Core/Representations/AbstractContentNegotiationPolicy.php
@@ -3,6 +3,8 @@ namespace BOTK\Core\Representations;
 
 use Bitworking\Mimeparse;
 use BOTK\Core\WebLink;
+use BOTK\Core\Http;
+
 
 /*
  * Exposes an index function that returns an array ('mime-type'=> rendering
@@ -114,5 +116,144 @@ abstract class AbstractContentNegotiationPolicy
         }
     }
 
+
+    /*************************************************************************
+     *  Default Serializers
+     *************************************************************************/
+
+
+    /**
+     * Serializ an Web link in htm head section
+     * 
+     * @link http://datatracker.ietf.org/doc/rfc4287 chapter 4.2.7
+     */
+    public static function htmlWebLinkSerializer(\BOTK\Core\WebLink $link) {
+        $tag = '<link href="'.$link->href.'"';
+        foreach (array('rel', 'type','hreflang') as $attribute) {
+            if( $value = $link->$attribute) $tag .= " $attribute='$value'";
+        }
+        $tag .= ' />'; 
+        
+        return $tag;       
+    }
+
+
+    /**
+     * Helper to render a nice view of Web link as html fragment
+     * 
+     */
+    public static function htmlWebLinks(array $links) {
+        $html = "<dl>\n";
+         $alternateLinks = array();
+         foreach($links as $link ){
+            $linkName = $link->rel?$link->rel:'Link';
+            $href=$link->href;
+            if(0===strcasecmp($linkName,'alternate')){
+                $type = $link->type?$link->type:'unspecfified type';
+                // manage antipatter to make alternate format ccallable from browser
+                $glue=(false===strpos($href, '?'))?'?':'&amp';
+                $forceOutput=$glue.'_output='.urlencode($type);
+                $alternateLinks[] = "<a href='$href{$forceOutput}'>$type</a> ";
+            } else {
+                $html .= "<dt>$linkName:</dt><dd><a href='$href'>$href</a></dd>\n";
+            }
+         }
+         if (count($alternateLinks)){
+            $html .= "<dt>alternate:</dt><dd>".implode(', ',$alternateLinks)."</dd>\n"; 
+         }
+        $html .= "</dl>\n";
+        
+        return $html;
+    }
+    
+    
+    /**
+     * if $useCustomFields use an xml representation for fields.. Not yet standarized in html5
+     * 
+     * This render use html5 semantic tagging.
+     */
+    public static function htmlSerializer($data, 
+            $meta = null,
+            $title  = null,         
+            $header = null,
+            $footer = null,
+            $dataIsHtmlFragment=null) {
+        //set defaults
+        if (is_null($meta)) {$meta=array();}
+        if (is_null($title)) {$title= is_object($data)?get_class($data):gettype($data);}
+        if (is_null($header)) {$header="<h1>$title</h1>";}
+        if (is_null($footer)) {$footer='';}
+        if (is_null($dataIsHtmlFragment)) {$dataIsHtmlFragment=false;}
+         
+        // initilalize metadata
+               
+        if (is_string($meta) && ($meta=trim($meta))){
+            $metadata[] = "<link rel='stylesheet' type='text/css' href='$meta'/>";
+        } elseif( is_array($metadata)){
+            $metadata = $meta;
+        } else {
+        	$metadata = array();
+        }
+         
+               
+        // prepare hyperlinks
+        $links = Http::getResponseLinks();
+        foreach ($links as $link) {
+            $metadata[] = self::htmlWebLinkSerializer($link);
+        }
+        $navLinks = empty($links)?'':static::htmlWebLinks($links);
+        
+        // prepare head    
+        $head = implode("\n",$metadata);
+            
+       // Prepare data content
+        $text = is_scalar($data)
+            ? ((string)$data)
+            : var_export($data, true);          
+        $htmlDataRepresentation = $dataIsHtmlFragment?$text:htmlspecialchars($text);
+         
+        return is_null($meta)  // if meta is null render just data as html fragment
+            ?  $htmlDataRepresentation
+            : ( "<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset='utf-8'>
+        <title>$title</title>
+        $head
+    </head>
+    <body>
+        <header>
+            $header
+            <nav>
+               <div id=botkHyperlinks>
+                $navLinks
+               </div>
+            </nav>
+        </header>
+<!-- main tag contains a php html encoded representation of public data --> 
+<!-- the class property contains the name of the data model from whom public data are extracted -->       
+$htmlDataRepresentation
+        <footer>
+            $footer
+        </footer>
+    </body>
+</html>
+"
+            );
+    }
+
+
+    /**
+     * an super simple template engine!
+     * 
+     */
+    public static function templateSerializer($data, $template) {
+        if(ob_start()){
+            @require $template;
+            $result = ob_get_contents();
+            ob_end_clean();
+        }
+        return true;
+    }    
 
 }